third_party/WebKit/LayoutTests/fast/dom/custom/lifecycle-created-createElement-recursion-expected.txt

CONSOLE ERROR: line 8: Uncaught RangeError: Maximum call stack size exceeded.
PASS unless crash
PASS successfullyParsed is true

TEST COMPLETE
